requirements.txt
annotated-types
anyio
click
colorama
fastapi
greenlet
h11
httptool
idna
psycopg2
pydantic
pydantic_core
python-dotenv
PyYAML
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
watchfiles
websockets
passlib[bcrypt]
bcrypt==3.2.0
pydantic-settings
python-jose[cryptography]
python-multipart # to process the Form Data ex: /login endpoint